requirements.txt
astroid==1.6.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.8.4
requests==2.18.4
six==1.11.0
todoist-python==7.0.18
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
